requirements.txt
Flask
langchain
openai
opencc-python-reimplemented
tiktoken
langchain-openai
langchain-community
chromadb
langchain-chroma
